+++ lenskit/lenskit/algorithms/basic.py
@@ -10,7 +10,6 @@ Basic utility algorithms and combiners.
 
 import logging
 from collections.abc import Iterable, Sequence
-from datetime import datetime
 from typing import overload
 
 import numpy as np
@@ -42,7 +41,7 @@ class PopScore(Predictor):
             - ``'count'``
 
     Attributes:
-        item_scores_(pandas.Series):
+        item_pop_(pandas.Series):
             Item popularity scores.
     """
 
@@ -52,13 +51,8 @@ class PopScore(Predictor):
     @override
     def fit(self, data: Dataset, **kwargs):
         _logger.info("counting item popularity")
-
-        counts = data.item_stats()["count"]
-        self.item_scores_ = self._fit_internal(counts, **kwargs)
-
-        return self
-
-    def _fit_internal(self, scores: pd.Series, **kwargs):
+        stats = data.item_stats()
+        scores = stats["count"]
         if self.score_method == "rank":
             _logger.info("ranking %d items", len(scores))
             scores = scores.rank().sort_index()
@@ -74,7 +68,9 @@ class PopScore(Predictor):
         else:
             raise ValueError("invalid scoring method " + repr(self.score_method))
 
-        return scores
+        self.item_scores_ = scores
+
+        return self
 
     @override
     def predict_for_user(self, user, items, ratings=None):
@@ -82,60 +78,6 @@ class PopScore(Predictor):
 
     def __str__(self):
         return "PopScore({})".format(self.score_method)
-
-
-class TimeBoundedPopScore(PopScore):
-    """
-    Score items by their time-bounded popularity, i.e., the popularity in the
-    most recent `time_window` period.  Use with :py:class:`TopN` to get a
-    most-popular-recent-items recommender.
-
-    Args:
-        time_window(datetime.timedelta):
-            The time window for computing popularity scores.
-        score_type(str):
-            The method for computing popularity scores.  Can be one of the following:
-
-            - ``'quantile'`` (the default)
-            - ``'rank'``
-            - ``'count'``
-
-    Attributes:
-        item_scores_(pandas.Series):
-            Time-bounded item popularity scores.
-    """
-
-    def __init__(self, cutoff: datetime, score_method="quantile"):
-        super().__init__(score_method)
-
-        self.cutoff = cutoff
-        self.score_method = score_method
-
-    @override
-    def fit(self, data: Dataset, **kwargs):
-        _logger.info("counting time-bounded item popularity")
-
-        log = data.interaction_log("numpy")
-
-        item_scores = None
-        if log.timestamps is None:
-            _logger.warning("no timestamps in interaction log; falling back to PopScore")
-            item_scores = super().fit(data, **kwargs).item_scores_
-        else:
-            counts = np.zeros(data.item_count, dtype=np.int32)
-            start_timestamp = self.cutoff.timestamp()
-            item_nums = log.item_nums[log.timestamps > start_timestamp]
-            np.add.at(counts, item_nums, 1)
-
-            item_scores = super()._fit_internal(pd.Series(counts, index=data.items.index), **kwargs)
-
-        self.item_scores_ = item_scores
-
-        return self
-
-    @override
-    def __str__(self):
-        return "TimeBoundedPopScore({}, {})".format(self.cutoff, self.score_method)
 
 
 class Memorized(Predictor):
